Move Provisioner class
